backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-pinecone>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
pypdf>=3.17.0
pinecone-client>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sentence-transformers>=2.2.0
